+++ checkMate/AppDelegate.swift
@@ -9,6 +9,8 @@ import GoogleSignIn
 import UIKit
 import Firebase
 
+import AuthenticationServices
+
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
@@ -24,6 +26,24 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         FirebaseApp.configure()
     
         signOut()
+        
+        /// 🍎 Apple login
+        let appleIDProvider = ASAuthorizationAppleIDProvider()
+        appleIDProvider.getCredentialState(forUserID: KeychainItem.currentUserIdentifier) { (credentialState, error) in
+            switch credentialState {
+            case .authorized:
+                break // The Apple ID credential is valid.
+            case .revoked, .notFound:
+                // The Apple ID credential is either revoked or was not found, so show the sign-in UI.
+                DispatchQueue.main.async {
+//                    self.window?.rootViewController?.showLoginViewController()
+                }
+            default:
+                break
+            }
+        }
+
+        
         return true
     }
